Finalizing changes

+++ user/backend/permission_backend/delete_permission.php
@@ -1,5 +1,8 @@
 <?php
 include '../../backend/config.php';
+include '../access.php';
+check_user_permission($allowed_permission_type, 'permission');
+checkUserPermission($allowed_permission_slug,'permission_delete');
 $per_id=$_GET['id'];
     $sql="DELETE FROM permission WHERE per_id=$per_id";
     $result=mysqli_query($connect,$sql);
